Refactor: Outsource STC event handler to its own module
- Moved STC logic into a dedicated file (stc.js)
- Added full JSDoc documentation for clarity and maintainability
- Replaced var with const for modern scoping
